+++ ui/dashboards/src/components/Variables/index.tsx
@@ -11,5 +11,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './components';
-export * from './core';
+export * from './Variable';
+export * from './VariableList';
