+++ ui/dashboards/src/components/index.ts
@@ -15,3 +15,4 @@ export * from './Dashboard';
 export * from './GridLayout';
 export * from './Panel';
 export * from './TimeRangeControls';
+export * from './Variables';
