+++ ui/dashboards/src/index.ts
@@ -14,4 +14,3 @@
 export * from './components';
 export * from './context';
 export * from './views';
-export * from './templating';
